+++ .editorconfig
@@ -2,3 +2,8 @@ root = true
 
 [*]
 end_of_line = lf
+insert_final_newline = true
+charset = utf-8
+indent_style = space
+indent_size = 2
+trim_trailing_whitespace = true
