+++ README.md
@@ -1,2 +1,20 @@
-# english-for-kids
-english-for-kids
+Vanilla JS project: **English for Kids**.
+
+#### Project deploy link [English for Kids](https://yan-black-english-for-kids.netlify.app/).
+
+## Implemented features
+* Game/Study mode
+* Statistics accumulation
+* Saving statistics in the localStorage
+* Flip cards
+* Responsibility
+
+### Availible Scripts
+* npm install (install all necessary dependencies)
+* npm run start (starts project at local server)
+* npm run build (builds a production version)
+* npm run lint (runs eslint at src/ folder)
+* npm run lint:fix (fix all fixabled problems)
+
+### Code quality
+* Eslint (Air-bnb style guide)
