+++ public/index.html
@@ -4,7 +4,6 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>English-for-Kids</title>
-    <link rel="stylesheet" href="src/style.css">
     <link rel="shortcut icon" href="src/assets//images/favicon.ico" type="image/x-icon">
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;500;700;900&display=swap" rel="stylesheet">
 </head>
@@ -78,6 +77,5 @@
             </a>
         </div>
     </div>
-    <script src="./dist/bundle.js" type="module"></script>
 </body>
 </html>
